Moved frame_allocator into memory directory
